Allow continuous 8,000 to 96,000 Hz host sample rates ('rate')
Although going above 48 kHz is rather pointless for normal usage,
FluidSynth supports up to 96 khz sample rates, so we might as well allow
it.

Modern OSes let the sample rate to be freely chosen, and running at very
specific rates allows for rendering WAV files at a given card's native
frequency without resampling; this might be useful for analysis purposes
(e.g., 49,716 Hz for OPL, 45,454 Hz for the Sound Blaster, or 32,000 Hz
for the MT-32).

+++ src/hardware/adlib_gold.cpp
@@ -30,7 +30,7 @@ CHECK_NARROWING();
 // Yamaha YM7128B Surround Processor emulation
 // -------------------------------------------
 
-SurroundProcessor::SurroundProcessor(const uint16_t sample_rate_hz)
+SurroundProcessor::SurroundProcessor(const uint32_t sample_rate_hz)
 {
 	assert(sample_rate_hz >= 10);
 
@@ -104,7 +104,7 @@ AudioFrame SurroundProcessor::Process(const AudioFrame frame)
 // Philips Semiconductors TDA8425 hi-fi stereo audio processor emulation
 // ---------------------------------------------------------------------
 
-StereoProcessor::StereoProcessor(const uint16_t _sample_rate_hz)
+StereoProcessor::StereoProcessor(const uint32_t _sample_rate_hz)
         : sample_rate_hz(_sample_rate_hz)
 {
 	assert(sample_rate_hz > 0);
@@ -326,7 +326,7 @@ AudioFrame StereoProcessor::Process(const AudioFrame frame)
 // AdLib Gold module
 // -----------------
 
-AdlibGold::AdlibGold(const uint16_t sample_rate_hz)
+AdlibGold::AdlibGold(const uint32_t sample_rate_hz)
         : surround_processor(nullptr),
           stereo_processor(nullptr)
 {
