+++ src/hardware/compressor.cpp
@@ -37,7 +37,7 @@ Compressor::Compressor() = default;
 
 Compressor::~Compressor() = default;
 
-void Compressor::Configure(const uint16_t _sample_rate_hz,
+void Compressor::Configure(const uint32_t _sample_rate_hz,
                            const float _0dbfs_sample_value, const float threshold_db,
                            const float _ratio, const float attack_time_ms,
                            const float release_time_ms, const float rms_window_ms)
